got the logic working for blockqoutes and made sure to move some of the logic around
